+++ packages/components-js/tests/e2e/specs/headline.e2e.ts
@@ -1,5 +1,6 @@
 import {
   getLifecycleStatus,
+  getProperty,
   selectNode,
   setContentWithDesignSystem,
   setProperty,
@@ -29,9 +30,27 @@ describe('headline', () => {
   };
 
   const getHost = () => selectNode(page, 'p-headline');
+  const getText = () => selectNode(page, 'p-headline >>> p-text');
 
   const getHeadlineTagName = async () =>
     await page.$eval('p-headline', (el) => el.shadowRoot.querySelector('.root').tagName);
+
+  it('should forward props correctly to p-text', async () => {
+    await initHeadline({ variant: 'inherit' });
+    const host = await getHost();
+    const text = await getText();
+
+    expect(await getProperty(text, 'size')).toBe('inherit');
+    expect(await getProperty(text, 'align')).toBe('left');
+    expect(await getProperty(text, 'ellipsis')).toBe(false);
+
+    await setProperty(host, 'align', 'center');
+    await setProperty(host, 'ellipsis', true);
+    await waitForStencilLifecycle(page);
+
+    expect(await getProperty(text, 'align')).toBe('center');
+    expect(await getProperty(text, 'ellipsis')).toBe(true);
+  });
 
   describe('tag', () => {
     it('should render according to variant', async () => {