+++ sketch/basic/CHANGELOG.md
@@ -8,6 +8,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [Unreleased]
 
+### [2.2.1] - 2021-11-16
+
+#### Changed
+
+Type | Basic Library |
+|---|---|
+Icon | Color variables |
+
+### [2.2.0] - 2021-11-12
+
+#### Added
+
+Type | Basic Library |
+|---|---|
+Icon | Success icon added
+
 ### [2.1.0] - 2021-06-08
 
 #### Added